Convert C code from tabs to spaces

+++ engine/libmkb/autogen.h
@@ -146,37 +146,37 @@ static const char * const opcodes[MK_OPCODES_LEN];
 #define MK_CORE_VOC_LEN (114)
 #define MK_VOC_ITEM_NAME_LEN (16)
 typedef struct mk_voc_item {
-	const char * const name[MK_VOC_ITEM_NAME_LEN];
-	const uint8_t type_code;
-	const u32 value;
+    const char * const name[MK_VOC_ITEM_NAME_LEN];
+    const uint8_t type_code;
+    const u32 value;
 } mk_voc_item_t;
 static const mk_voc_item_t mk_core_voc[MK_CORE_VOC_LEN];
 
 // VM context struct for holding state of registers and RAM
 #define MK_BufMax (256)
 typedef struct mk_context {
-	u8  err;               // Error register (don't confuse with ERR opcode!)
-	u8  base;              // number Base for debug printing
-	i32 A;                 // register for source address or scratch
-	i32 B;                 // register for destination addr or scratch
-	i32 T;                 // Top of data stack
-	i32 S;                 // Second on data stack
-	i32 R;                 // top of Return stack
-	u16 PC;                // Program Counter
-	u8  DSDeep;            // Data Stack Depth (count include T and S)
-	u8  RSDeep;            // Return Stack Depth (count inlcudes R)
-	i32 DStack[16];        // Data Stack
-	i32 RStack[16];        // Return Stack
-	u8  RAM[MK_MemMax+1];  // Random Access Memory
-	u8  InBuf[MK_BufMax];  // Input buffer
-	u8  OutBuf[MK_BufMax]; // Output buffer
-	u8  echo;              // Echo depends on tty vs pip, etc.
-	u8  halted;            // Flag to track halt (used for `bye`)
-	u8  HoldStdout;        // Flag to use holding buffer for stdout
-	u8  IOLOAD_depth;      // Nesting level for io_load_file()
-	u8  IOLOAD_fail;       // Flag indicating an error during io_load_file()
-	u8  FOPEN_file;        // File (if any) that was opened by FOPEN
-	u8  DbgTraceEnable;    // Debug trace on/off
+    u8  err;               // Error register (don't confuse with ERR opcode!)
+    u8  base;              // number Base for debug printing
+    i32 A;                 // register for source address or scratch
+    i32 B;                 // register for destination addr or scratch
+    i32 T;                 // Top of data stack
+    i32 S;                 // Second on data stack
+    i32 R;                 // top of Return stack
+    u32 PC;                // Program Counter
+    u8  DSDeep;            // Data Stack Depth (count include T and S)
+    u8  RSDeep;            // Return Stack Depth (count inlcudes R)
+    i32 DStack[16];        // Data Stack
+    i32 RStack[16];        // Return Stack
+    u8  RAM[MK_MemMax+1];  // Random Access Memory
+    u8  InBuf[MK_BufMax];  // Input buffer
+    u8  OutBuf[MK_BufMax]; // Output buffer
+    u8  echo;              // Echo depends on tty vs pip, etc.
+    u8  halted;            // Flag to track halt (used for `bye`)
+    u8  HoldStdout;        // Flag to use holding buffer for stdout
+    u8  IOLOAD_depth;      // Nesting level for io_load_file()
+    u8  IOLOAD_fail;       // Flag indicating an error during io_load_file()
+    u8  FOPEN_file;        // File (if any) that was opened by FOPEN 
+    u8  DbgTraceEnable;    // Debug trace on/off
 } mk_context_t;
 
 // Maximum number of cycles allowed before infinite loop error triggers
